requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
redis[asyncio]==5.2.1
python-dotenv==1.1.0
fastapi-users[sqlalchemy]==14.0.1
aiosqlite==0.21.0
